Split impulse-form-shape.spec
