+++ packages/react-impulse-form/tests/impulse-form-shape/set-error.spec.ts
@@ -1,0 +1,200 @@
+import type { Setter } from "~/tools/setter"
+
+import { ImpulseFormShape, ImpulseFormUnit } from "../../src"
+import { arg } from "../common"
+
+it("specifies error", ({ scope }) => {
+  const shape = ImpulseFormShape({
+    first: ImpulseFormUnit("", { error: ["first"] }),
+    second: ImpulseFormUnit(0, { error: ["second"] }),
+    third: ImpulseFormShape(
+      {
+        one: ImpulseFormUnit(true, {
+          validate: (input) => (input ? [null, input] : ["must be true", null]),
+        }),
+        two: ImpulseFormUnit([""], { error: "an error" }),
+      },
+      {
+        error: {
+          one: "one",
+          two: "two",
+        },
+      },
+    ),
+    fourth: ["anything"],
+  })
+
+  expect(shape.getError(scope)).toStrictEqual({
+    first: ["first"],
+    second: ["second"],
+    third: {
+      one: "one",
+      two: "two",
+    },
+  })
+
+  shape.setError({
+    first: ["another"],
+    second: undefined,
+    third: null,
+  })
+  expect(shape.getError(scope)).toStrictEqual({
+    first: ["another"],
+    second: ["second"],
+    third: null,
+  })
+  expect(shape.getError(scope, arg(1))).toStrictEqual({
+    first: ["another"],
+    second: ["second"],
+    third: {
+      one: null,
+      two: null,
+    },
+  })
+
+  shape.setError({
+    third: {
+      one: "one",
+      two: "two",
+    },
+  })
+  shape.setError((root) => {
+    expectTypeOf(root).toEqualTypeOf<{
+      readonly first: null | Array<string>
+      readonly second: null | Array<string>
+      readonly third: {
+        readonly one: null | string
+        readonly two: null | string
+      }
+    }>()
+    expect(root).toStrictEqual({
+      first: ["another"],
+      second: ["second"],
+      third: {
+        one: "one",
+        two: "two",
+      },
+    })
+
+    return {
+      first: (first) => {
+        expectTypeOf(first).toEqualTypeOf<null | Array<string>>()
+        expect(first).toStrictEqual(["another"])
+
+        return [...first!, "1"]
+      },
+      second: (second) => {
+        expectTypeOf(second).toEqualTypeOf<null | Array<string>>()
+        expect(second).toStrictEqual(["second"])
+
+        return [...second!, "2"]
+      },
+      third: (third) => {
+        expectTypeOf(third).toEqualTypeOf<{
+          readonly one: null | string
+          readonly two: null | string
+        }>()
+        expect(third).toStrictEqual({
+          one: "one",
+          two: "two",
+        })
+
+        return {
+          one: (one) => {
+            expectTypeOf(one).toEqualTypeOf<null | string>()
+            expect(one).toStrictEqual("one")
+
+            return "1"
+          },
+
+          two: (two) => {
+            expectTypeOf(two).toEqualTypeOf<null | string>()
+            expect(two).toStrictEqual("two")
+
+            return "2"
+          },
+        }
+      },
+    }
+  })
+
+  expect(shape.getError(scope)).toStrictEqual({
+    first: ["another", "1"],
+    second: ["second", "2"],
+    third: {
+      one: "1",
+      two: "2",
+    },
+  })
+
+  expectTypeOf(shape.setError).parameter(0).toEqualTypeOf<
+    Setter<
+      null | {
+        readonly first?: Setter<null | Array<string>>
+        readonly second?: Setter<null | Array<string>>
+        readonly third?: Setter<
+          null | {
+            readonly one?: Setter<null | string>
+            readonly two?: Setter<null | string>
+          },
+          [
+            {
+              readonly one: null | string
+              readonly two: null | string
+            },
+          ]
+        >
+      },
+      [
+        {
+          readonly first: null | Array<string>
+          readonly second: null | Array<string>
+          readonly third: {
+            readonly one: null | string
+            readonly two: null | string
+          }
+        },
+      ]
+    >
+  >()
+
+  expectTypeOf(shape.fields.third.setError).parameter(0).toEqualTypeOf<
+    Setter<
+      null | {
+        readonly one?: Setter<null | string>
+        readonly two?: Setter<null | string>
+      },
+      [
+        {
+          readonly one: null | string
+          readonly two: null | string
+        },
+      ]
+    >
+  >()
+})
+
+it("resets all errors", ({ scope }) => {
+  const shape = ImpulseFormShape({
+    first: ImpulseFormUnit("", { error: ["first"] }),
+    second: ImpulseFormUnit(0, { error: ["second"] }),
+    third: ImpulseFormShape(
+      {
+        one: ImpulseFormUnit(true, {
+          validate: (input) => (input ? [null, input] : [1, null]),
+        }),
+        two: ImpulseFormUnit([""], { error: "an error" }),
+      },
+      {
+        error: {
+          one: 0,
+          two: "initial error",
+        },
+      },
+    ),
+    fourth: ["anything"],
+  })
+
+  shape.setError(null)
+  expect(shape.getError(scope)).toBeNull()
+})
